Moved sortList and filterCategoryList into filterSortHandler.js
